+++ src/protocol/oprf/mod.rs
@@ -1,157 +1,21 @@
+pub mod oprf_share;
+
 use std::ops::Add;
 
 use futures::future;
-use generic_array::GenericArray;
 use ipa_macros::Step;
 use rand::{distributions::Standard, seq::SliceRandom, Rng};
-use typenum::Unsigned;
 
-use super::{context::Context, ipa::IPAInputRow, RecordId};
+use self::oprf_share::{OPRFShare, OprfBK, OprfF, OprfMK};
+use super::{
+    context::Context, ipa::IPAInputRow, sort::apply::apply as apply_permutation, RecordId,
+};
 use crate::{
     error::Error,
-    ff::{Field, Gf32Bit, Gf40Bit, Gf8Bit, Serializable},
-    helpers::{query::oprf_shuffle::QueryConfig, Direction, Message, ReceivingEnd, Role},
+    helpers::{query::oprf_shuffle::QueryConfig, Direction, ReceivingEnd, Role},
 };
 
-type OprfMK = Gf40Bit;
-type OprfBK = Gf8Bit;
-type OprfF = Gf32Bit;
-
 pub type OPRFInputRow = IPAInputRow<OprfF, OprfMK, OprfBK>;
-
-#[derive(Debug, Clone, Copy)]
-pub struct OPRFShuffleSingleShare {
-    pub timestamp: OprfF,
-    pub mk: OprfMK,
-    pub is_trigger_bit: OprfF,
-    pub breakdown_key: OprfBK,
-    pub trigger_value: OprfF,
-}
-
-impl OPRFShuffleSingleShare {
-    #[must_use]
-    pub fn from_input_row(input_row: &OPRFInputRow, shared_with: Direction) -> Self {
-        match shared_with {
-            Direction::Left => Self {
-                timestamp: input_row.timestamp.as_tuple().1,
-                mk: input_row.mk_shares.as_tuple().1,
-                is_trigger_bit: input_row.is_trigger_bit.as_tuple().1,
-                breakdown_key: input_row.breakdown_key.as_tuple().1,
-                trigger_value: input_row.trigger_value.as_tuple().1,
-            },
-
-            Direction::Right => Self {
-                timestamp: input_row.timestamp.as_tuple().0,
-                mk: input_row.mk_shares.as_tuple().0,
-                is_trigger_bit: input_row.is_trigger_bit.as_tuple().0,
-                breakdown_key: input_row.breakdown_key.as_tuple().0,
-                trigger_value: input_row.trigger_value.as_tuple().0,
-            },
-        }
-    }
-
-    #[must_use]
-    pub fn to_input_row(self, rhs: Self) -> OPRFInputRow {
-        OPRFInputRow {
-            timestamp: (self.timestamp, rhs.timestamp).into(),
-            mk_shares: (self.mk, rhs.mk).into(),
-            is_trigger_bit: (self.is_trigger_bit, rhs.is_trigger_bit).into(),
-            breakdown_key: (self.breakdown_key, rhs.breakdown_key).into(),
-            trigger_value: (self.trigger_value, rhs.trigger_value).into(),
-        }
-    }
-}
-
-impl rand::prelude::Distribution<OPRFShuffleSingleShare> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> OPRFShuffleSingleShare {
-        OPRFShuffleSingleShare {
-            timestamp: OprfF::truncate_from(rng.gen::<u128>()),
-            mk: OprfMK::truncate_from(rng.gen::<u128>()),
-            is_trigger_bit: OprfF::truncate_from(rng.gen::<u128>()),
-            breakdown_key: OprfBK::truncate_from(rng.gen::<u128>()),
-            trigger_value: OprfF::truncate_from(rng.gen::<u128>()),
-        }
-    }
-}
-
-impl Add for OPRFShuffleSingleShare {
-    type Output = Self;
-
-    fn add(self, rhs: Self) -> Self::Output {
-        Add::add(&self, &rhs)
-    }
-}
-
-impl<'a, 'b> Add<&'b OPRFShuffleSingleShare> for &'a OPRFShuffleSingleShare {
-    type Output = OPRFShuffleSingleShare;
-
-    fn add(self, rhs: &'b OPRFShuffleSingleShare) -> Self::Output {
-        Self::Output {
-            timestamp: self.timestamp + rhs.timestamp,
-            mk: self.mk + rhs.mk,
-            is_trigger_bit: self.is_trigger_bit + rhs.is_trigger_bit,
-            breakdown_key: self.breakdown_key + rhs.breakdown_key,
-            trigger_value: self.trigger_value + rhs.trigger_value,
-        }
-    }
-}
-
-impl Serializable for OPRFShuffleSingleShare {
-    type Size = <<OprfF as Serializable>::Size as Add<
-        <<OprfMK as Serializable>::Size as Add<
-            <<OprfF as Serializable>::Size as Add<
-                <<OprfBK as Serializable>::Size as Add<<OprfF as Serializable>::Size>>::Output,
-            >>::Output,
-        >>::Output,
-    >>::Output;
-
-    fn serialize(&self, buf: &mut GenericArray<u8, Self::Size>) {
-        let mk_sz = <OprfMK as Serializable>::Size::USIZE;
-        let bk_sz = <OprfBK as Serializable>::Size::USIZE;
-        let f_sz = <OprfF as Serializable>::Size::USIZE;
-
-        self.timestamp
-            .serialize(GenericArray::from_mut_slice(&mut buf[..f_sz]));
-        self.mk
-            .serialize(GenericArray::from_mut_slice(&mut buf[f_sz..f_sz + mk_sz]));
-        self.is_trigger_bit.serialize(GenericArray::from_mut_slice(
-            &mut buf[f_sz + mk_sz..f_sz + mk_sz + f_sz],
-        ));
-        self.breakdown_key.serialize(GenericArray::from_mut_slice(
-            &mut buf[f_sz + mk_sz + f_sz..f_sz + mk_sz + f_sz + bk_sz],
-        ));
-        self.trigger_value.serialize(GenericArray::from_mut_slice(
-            &mut buf[f_sz + mk_sz + f_sz + bk_sz..],
-        ));
-    }
-
-    fn deserialize(buf: &GenericArray<u8, Self::Size>) -> Self {
-        let mk_sz = <OprfMK as Serializable>::Size::USIZE;
-        let bk_sz = <OprfBK as Serializable>::Size::USIZE;
-        let f_sz = <OprfF as Serializable>::Size::USIZE;
-
-        let timestamp = OprfF::deserialize(GenericArray::from_slice(&buf[..f_sz]));
-        let mk = OprfMK::deserialize(GenericArray::from_slice(&buf[f_sz..f_sz + mk_sz]));
-        let is_trigger_bit = OprfF::deserialize(GenericArray::from_slice(
-            &buf[f_sz + mk_sz..f_sz + mk_sz + f_sz],
-        ));
-        let breakdown_key = OprfBK::deserialize(GenericArray::from_slice(
-            &buf[f_sz + mk_sz + f_sz..f_sz + mk_sz + f_sz + bk_sz],
-        ));
-        let trigger_value = OprfF::deserialize(GenericArray::from_slice(
-            &buf[f_sz + mk_sz + f_sz + bk_sz..],
-        ));
-        Self {
-            timestamp,
-            mk,
-            is_trigger_bit,
-            breakdown_key,
-            trigger_value,
-        }
-    }
-}
-
-impl Message for OPRFShuffleSingleShare {}
 
 #[derive(Step)]
 pub(crate) enum OPRFShuffleStep {
@@ -206,10 +70,10 @@ async fn run_h1<C, Sl, Sr, Zl, Zr>(
 ) -> Result<Vec<OPRFInputRow>, Error>
 where
     C: Context,
-    Sl: IntoIterator<Item = OPRFShuffleSingleShare>,
-    Sr: IntoIterator<Item = OPRFShuffleSingleShare>,
-    Zl: IntoIterator<Item = OPRFShuffleSingleShare>,
-    Zr: IntoIterator<Item = OPRFShuffleSingleShare>,
+    Sl: IntoIterator<Item = OPRFShare>,
+    Sr: IntoIterator<Item = OPRFShare>,
+    Zl: IntoIterator<Item = OPRFShare>,
+    Zr: IntoIterator<Item = OPRFShare>,
 {
     // 1. Generate helper-specific random tables
     let ctx_a_hat = ctx.narrow(&OPRFShuffleStep::GenerateAHat);
@@ -221,12 +85,11 @@ where
         generate_random_table_solo(batch_size, &ctx_b_hat, Direction::Right).collect();
 
     // 2. Run computations
-    let mut x_1: Vec<OPRFShuffleSingleShare> =
-        add_single_shares(add_single_shares(a, b), z_12).collect();
-    apply(&pi_12, &mut x_1);
+    let mut x_1: Vec<OPRFShare> = add_single_shares(add_single_shares(a, b), z_12).collect();
+    apply_permutation(&pi_12, &mut x_1);
 
-    let mut x_2: Vec<OPRFShuffleSingleShare> = add_single_shares(x_1, z_31).collect();
-    apply(&pi_31, &mut x_2);
+    let mut x_2: Vec<OPRFShare> = add_single_shares(x_1, z_31).collect();
+    apply_permutation(&pi_31, &mut x_2);
 
     send_to_peer(ctx, &OPRFShuffleStep::TransferX2, Direction::Right, x_2).await?;
 
@@ -243,10 +106,10 @@ async fn run_h2<C, Sl, Sr, Zl, Zr>(
 ) -> Result<Vec<OPRFInputRow>, Error>
 where
     C: Context,
-    Sl: IntoIterator<Item = OPRFShuffleSingleShare>,
-    Sr: IntoIterator<Item = OPRFShuffleSingleShare>,
-    Zl: IntoIterator<Item = OPRFShuffleSingleShare>,
-    Zr: IntoIterator<Item = OPRFShuffleSingleShare>,
+    Sl: IntoIterator<Item = OPRFShare>,
+    Sr: IntoIterator<Item = OPRFShare>,
+    Zl: IntoIterator<Item = OPRFShare>,
+    Zr: IntoIterator<Item = OPRFShare>,
 {
     // 1. Generate helper-specific random tables
     let ctx_b_hat = ctx.narrow(&OPRFShuffleStep::GenerateBHat);
@@ -254,8 +117,8 @@ where
         generate_random_table_solo(batch_size, &ctx_b_hat, Direction::Left).collect();
 
     // 2. Run computations
-    let mut y_1: Vec<OPRFShuffleSingleShare> = add_single_shares(c, z_12.into_iter()).collect();
-    apply(&pi_12, &mut y_1);
+    let mut y_1: Vec<OPRFShare> = add_single_shares(c, z_12.into_iter()).collect();
+    apply_permutation(&pi_12, &mut y_1);
 
     let ((), x_2) = future::try_join(
         send_to_peer(ctx, &OPRFShuffleStep::TransferY1, Direction::Right, y_1),
@@ -269,7 +132,7 @@ where
     .await?;
 
     let mut x_3: Vec<_> = add_single_shares(x_2.into_iter(), z_23.into_iter()).collect();
-    apply(&pi_23, &mut x_3);
+    apply_permutation(&pi_23, &mut x_3);
 
     let c_hat_1: Vec<_> = add_single_shares(x_3.iter(), b_hat.iter()).collect();
     let ((), c_hat_2) = future::try_join(
@@ -302,10 +165,10 @@ async fn run_h3<C, Sl, Sr, Zl, Zr>(
 ) -> Result<Vec<OPRFInputRow>, Error>
 where
     C: Context,
-    Sl: IntoIterator<Item = OPRFShuffleSingleShare>,
-    Sr: IntoIterator<Item = OPRFShuffleSingleShare>,
-    Zl: IntoIterator<Item = OPRFShuffleSingleShare>,
-    Zr: IntoIterator<Item = OPRFShuffleSingleShare>,
+    Sl: IntoIterator<Item = OPRFShare>,
+    Sr: IntoIterator<Item = OPRFShare>,
+    Zl: IntoIterator<Item = OPRFShare>,
+    Zr: IntoIterator<Item = OPRFShare>,
 {
     // 1. Generate helper-specific random tables
     let ctx_a_hat = ctx.narrow(&OPRFShuffleStep::GenerateAHat);
@@ -321,11 +184,11 @@ where
     )
     .await?;
 
-    let mut y_2: Vec<OPRFShuffleSingleShare> = add_single_shares(y_1, z_31).collect();
-    apply(&pi_31, &mut y_2);
+    let mut y_2: Vec<OPRFShare> = add_single_shares(y_1, z_31).collect();
+    apply_permutation(&pi_31, &mut y_2);
 
-    let mut y_3: Vec<OPRFShuffleSingleShare> = add_single_shares(y_2, z_23).collect();
-    apply(&pi_23, &mut y_3);
+    let mut y_3: Vec<OPRFShare> = add_single_shares(y_2, z_23).collect();
+    apply_permutation(&pi_23, &mut y_3);
 
     let c_hat_2 = add_single_shares(y_3.iter(), a_hat.iter()).collect::<Vec<_>>();
     let ((), c_hat_1) = future::try_join(
@@ -354,15 +217,15 @@ where
 fn split_shares(
     input_rows: &[OPRFInputRow],
     direction: Direction,
-) -> impl Iterator<Item = OPRFShuffleSingleShare> + '_ {
-    let f = move |input_row| OPRFShuffleSingleShare::from_input_row(input_row, direction);
+) -> impl Iterator<Item = OPRFShare> + '_ {
+    let f = move |input_row| OPRFShare::from_input_row(input_row, direction);
     input_rows.iter().map(f)
 }
 
 fn combine_shares<L, R>(l: L, r: R) -> Vec<OPRFInputRow>
 where
-    L: IntoIterator<Item = OPRFShuffleSingleShare>,
-    R: IntoIterator<Item = OPRFShuffleSingleShare>,
+    L: IntoIterator<Item = OPRFShare>,
+    R: IntoIterator<Item = OPRFShare>,
 {
     l.into_iter()
         .zip(r)
@@ -385,8 +248,8 @@ fn generate_random_tables_with_peers<C: Context>(
     batch_size: u32,
     narrow_ctx: &C,
 ) -> (
-    impl Iterator<Item = OPRFShuffleSingleShare> + '_,
-    impl Iterator<Item = OPRFShuffleSingleShare> + '_,
+    impl Iterator<Item = OPRFShare> + '_,
+    impl Iterator<Item = OPRFShare> + '_,
 ) {
     let (rng_l, rng_r) = narrow_ctx.prss_rng();
     let with_left = sample_iter(rng_l).take(batch_size as usize);
@@ -398,7 +261,7 @@ fn generate_random_table_solo<C>(
     batch_size: u32,
     narrow_ctx: &C,
     peer: Direction,
-) -> impl Iterator<Item = OPRFShuffleSingleShare> + '_
+) -> impl Iterator<Item = OPRFShare> + '_
 where
     C: Context,
 {
@@ -411,13 +274,13 @@ where
     sample_iter(rng).take(batch_size as usize)
 }
 
-fn sample_iter<R: Rng>(rng: R) -> impl Iterator<Item = OPRFShuffleSingleShare> {
+fn sample_iter<R: Rng>(rng: R) -> impl Iterator<Item = OPRFShare> {
     rng.sample_iter(Standard)
 }
 
 // ---------------------------- helper communication ------------------------------------ //
 
-async fn send_to_peer<C: Context, I: IntoIterator<Item = OPRFShuffleSingleShare>>(
+async fn send_to_peer<C: Context, I: IntoIterator<Item = OPRFShare>>(
     ctx: &C,
     step: &OPRFShuffleStep,
     direction: Direction,
@@ -436,11 +299,11 @@ async fn receive_from_peer<C: Context>(
     step: &OPRFShuffleStep,
     direction: Direction,
     batch_size: u32,
-) -> Result<Vec<OPRFShuffleSingleShare>, Error> {
+) -> Result<Vec<OPRFShare>, Error> {
     let role = ctx.role().peer(direction);
-    let receive_channel: ReceivingEnd<OPRFShuffleSingleShare> = ctx.narrow(step).recv_channel(role);
+    let receive_channel: ReceivingEnd<OPRFShare> = ctx.narrow(step).recv_channel(role);
 
-    let mut output: Vec<OPRFShuffleSingleShare> = Vec::with_capacity(batch_size as usize);
+    let mut output: Vec<OPRFShare> = Vec::with_capacity(batch_size as usize);
     for record_id in 0..batch_size {
         let msg = receive_channel.receive(RecordId::from(record_id)).await?;
         output.push(msg);
@@ -464,31 +327,4 @@ fn generate_pseudorandom_permutation<R: Rng>(batch_size: u32, rng: &mut R) -> Ve
     let mut permutation = (0..batch_size).collect::<Vec<_>>();
     permutation.shuffle(rng);
     permutation
-}
-
-use bitvec::bitvec;
-use embed_doc_image::embed_doc_image;
-
-#[embed_doc_image("apply", "images/sort/apply.png")]
-/// Permutation reorders (1, 2, . . . , m) into (σ(1), σ(2), . . . , σ(m)).
-/// For example, if σ(1) = 2, σ(2) = 3, σ(3) = 1, and σ(4) = 0, an input (A, B, C, D) is reordered into (C, D, B, A) by σ.
-///
-/// ![Apply steps][apply]
-fn apply<T>(permutation: &[u32], values: &mut [T]) {
-    // NOTE: This is copypasta from crate::protocol::sort
-    debug_assert!(permutation.len() == values.len());
-    let mut permuted = bitvec![0; permutation.len()];
-
-    for i in 0..permutation.len() {
-        if !permuted[i] {
-            let mut pos_i = i;
-            let mut pos_j = permutation[pos_i] as usize;
-            while pos_j != i {
-                values.swap(pos_i, pos_j);
-                permuted.set(pos_j, true);
-                pos_i = pos_j;
-                pos_j = permutation[pos_i] as usize;
-            }
-        }
-    }
 }
